requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
avalanche-lib==0.6.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
clarabel==0.10.0
click==8.1.8
cmeel==0.57.1
contourpy==1.3.0
cvxopt==1.3.2
cycler==0.12.1
daqp==0.6.0
dill==0.3.9
docker-pycreds==0.4.0
ecos==2.0.14
eval_type_backport==0.2.2
fedlab==1.3.0
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
GPUtil==1.4.0
grpcio==1.71.0
highspy==1.9.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
jax==0.4.30
jaxlib==0.4.30
jaxopt==0.8.3
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.7
lightning-utilities==0.14.0
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.4
ml_dtypes==0.5.1
mpmath==1.3.0
munch==4.0.0
networkx==3.2.1
numpy==2.0.2
nvidia-ml-py==12.570.86
opt_einsum==3.4.0
osqp==0.6.7.post3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
piqp==0.4.2
platformdirs==4.3.6
protobuf==5.29.3
proxsuite==0.7.2
psutil==7.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pynvml==12.0.0
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytorchcv==0.0.67
pytz==2025.1
PyYAML==6.0.2
qdldl==0.1.7.post5
qpalm==1.2.5
qpax==0.0.9
qpsolvers==4.5.0
quadprog==0.1.13
requests==2.32.3
scikit-learn==1.6.1
scipy==1.13.1
scs==3.2.7.post2
sentry-sdk==2.22.0
setproctitle==1.3.5
six==1.17.0
smmap==5.0.2
soupsieve==2.6
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tomli==2.2.1
torch==2.6.0
torchmetrics==1.6.3
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wandb==0.19.8
Werkzeug==3.1.3
zipp==3.21.0
